backend/requirements.txt
Flask==2.3.2
pymodbus==3.3.1
python-dotenv==1.0.0
Flask-Cors==3.0.10